AIVLE_Backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
bcrypt==4.1.3
bidict==0.23.1
branca==0.7.2
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
eventlet==0.36.1
filelock==3.15.4
folium==0.17.0
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
install==1.3.5
intel-openmp==2021.4.0
Jinja2==3.1.4
jwt==1.3.1
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.35.13
packaging==24.1
pandas==2.2.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.13.0
tbb==2021.13.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wsproto==1.2.0
xyzservices==2024.6.0
